Framebuffer Objects and Normalized UV Coordinates
Added framebuffer objects, which will take one or more renderable
objects and render them to a texture. Also normalized UV coordinates by
loading images and flipping their Y axis (making the origin the bottom
left hand corner). Then made Sprites use a new normalize flipped y
function which means that all coordinates for the texture will be stored
with an origin at the top left, but change it to bottom left when
buffering the data to the GPU.