Added Cursor handlers and Cursor states
New cursor handler and example mouse/cursor handler which moves the
renderer's camera and disables the cursor when a mouse button is
pressed. Updated the window manager to add an updateWindows function to
perform all of the window updates that must happen in the main (GFWL
initialized) thread. Events also now store a reference to the window
they modify, instead of the window handle.